requirements.txt
gunicorn
flask
flask_httpauth
werkzeug
tensorflow
torch
numpy
pillow
opencv-python
requests
pytesseract
SpeechRecognition
moviepy
requests
googletrans